+++ src/Attributes/DisableQueryTracking.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Mattiasgeniar\PhpunitQueryCountAssertions\Attributes;
+
+use Attribute;
+
+#[Attribute(Attribute::TARGET_METHOD | Attribute::TARGET_CLASS)]
+class DisableQueryTracking {}